properly split UCAC5 source and library
